restructure the catalog pages in the example app
