+++ packages/app/src/components/catalog/kinds/SystemEntityPage.tsx
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2021 Spotify AB
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Entity, SystemEntity } from '@backstage/catalog-model';
+import { EntityHasApisCard } from '@backstage/plugin-api-docs';
+import {
+  AboutCard,
+  EntityHasComponentsCard,
+  EntityPageLayout,
+} from '@backstage/plugin-catalog';
+import { Grid } from '@material-ui/core';
+import React from 'react';
+
+const OverviewContent = ({ entity }: { entity: SystemEntity }) => (
+  <Grid container alignItems="stretch">
+    <Grid item md={6}>
+      <AboutCard entity={entity} variant="gridItem" />
+    </Grid>
+    <Grid item md={6}>
+      <EntityHasComponentsCard variant="gridItem" />
+    </Grid>
+    <Grid item md={6}>
+      <EntityHasApisCard variant="gridItem" />
+    </Grid>
+  </Grid>
+);
+
+export const SystemEntityPage = ({ entity }: { entity: Entity }) => (
+  <EntityPageLayout>
+    <EntityPageLayout.Content
+      path="/*"
+      title="Overview"
+      element={<OverviewContent entity={entity as SystemEntity} />}
+    />
+  </EntityPageLayout>
+);
